+++ src/Federation/Directives/ExternalDirective.php
@@ -3,7 +3,7 @@
 namespace Nuwave\Lighthouse\Federation\Directives;
 
 use GraphQL\Executor\Executor;
-use GraphQL\Type\Definition\ResolveInfo;
+use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;