+++ src/Schema/Directives/RelationDirectiveHelpers.php
@@ -2,9 +2,10 @@
 
 namespace Nuwave\Lighthouse\Schema\Directives;
 
-use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Relations\Relation;
+use Illuminate\Database\Query\Builder as QueryBuilder;
+use Nuwave\Lighthouse\Execution\ResolveInfo;
 
 trait RelationDirectiveHelpers
 {
@@ -26,7 +27,7 @@ trait RelationDirectiveHelpers
     }
 
     /**
-     * @return \Closure(\Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Eloquent\Relations\Relation): void
+     * @return \Closure(QueryBuilder|\Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Eloquent\Relations\Relation): void
      */
     protected function makeBuilderDecorator(ResolveInfo $resolveInfo): \Closure
     {
@@ -34,9 +35,9 @@ trait RelationDirectiveHelpers
             if ($builder instanceof Relation) {
                 $builder = $builder->getQuery();
             }
-            assert($builder instanceof EloquentBuilder);
+            assert($builder instanceof QueryBuilder || $builder instanceof EloquentBuilder);
 
-            $resolveInfo->argumentSet->enhanceBuilder(
+            $resolveInfo->enhanceBuilder(
                 $builder,
                 $this->scopes()
             );