+++ src/Testing/MakesGraphQLRequestsLumen.php
@@ -36,9 +36,9 @@ trait MakesGraphQLRequestsLumen
     protected $deferStream;
 
     /**
-     * Execute a query as if it was sent as a request to the server.
+     * Execute a GraphQL operation as if it was sent as a request to the server.
      *
-     * @param  string  $query  The GraphQL query to send
+     * @param  string  $query  The GraphQL operation to send
      * @param  array<string, mixed>  $variables  The variables to include in the query
      * @param  array<string, mixed>  $extraParams  Extra parameters to add to the JSON payload
      * @param  array<string, mixed>  $headers  HTTP headers to pass to the POST request
@@ -62,7 +62,7 @@ trait MakesGraphQLRequestsLumen
     }
 
     /**
-     * Execute a POST to the GraphQL endpoint.
+     * Send a POST request to the GraphQL endpoint.
      *
      * Use this over graphQL() when you need more control or want to
      * test how your server behaves on incorrect inputs.
@@ -82,7 +82,7 @@ trait MakesGraphQLRequestsLumen
     }
 
     /**
-     * Send a multipart form request to GraphQL.
+     * Send a multipart form request to the GraphQL endpoint.
      *
      * This is used for file uploads conforming to the specification:
      * https://github.com/jaydenseric/graphql-multipart-request-spec
@@ -123,7 +123,9 @@ trait MakesGraphQLRequestsLumen
     }
 
     /**
-     * Execute the introspection query on the GraphQL server.
+     * Send the introspection query to the GraphQL server.
+     *
+     * Returns the cached first result on repeated calls.
      */
     protected function introspect(): self
     {
@@ -233,12 +235,13 @@ trait MakesGraphQLRequestsLumen
         $this->deferStream = new MemoryStream();
 
         Container::getInstance()->singleton(CanStreamResponse::class, function (): MemoryStream {
-            assert($this->deferStream instanceof MemoryStream);
-
             return $this->deferStream;
         });
     }
 
+    /**
+     * Configure an error handler that rethrows all errors passed to it.
+     */
     protected function rethrowGraphQLErrors(): void
     {
         $config = Container::getInstance()->make(ConfigRepository::class);