+++ tests/Integration/Scout/SearchDirectiveTest.php
@@ -161,7 +161,7 @@ final class SearchDirectiveTest extends DBTestCase
     /**
      * @param  array{id: int}  $value
      */
-    public function customBuilderMethod(ScoutBuilder $builder, array $value): ScoutBuilder
+    public static function customBuilderMethod(ScoutBuilder $builder, array $value): ScoutBuilder
     {
         return $builder->where('from_custom_builder', $value['id']);
     }