+++ tests/Unit/Schema/Directives/Fixtures/Foo.php
@@ -6,7 +6,7 @@ namespace Tests\Unit\Schema\Directives\Fixtures;
  * TODO remove in favor of ->getMockBuilder(\stdClass::class)->addMethods(['__invoke'])
  * once we no longer support PHPUnit 7.
  */
-class Foo
+/* not final because mocked */ class Foo
 {
     public function bar(): void
     {