+++ tests/Unit/Tests/Unit/Auth/CanDirectiveTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Unit\Auth;
 
+use Nuwave\Lighthouse\Auth\CanDirective;
 use Nuwave\Lighthouse\Exceptions\AuthorizationException;
 use Nuwave\Lighthouse\Support\AppVersion;
 use Tests\TestCase;
@@ -346,11 +347,21 @@ final class CanDirectiveTest extends TestCase
         ]);
     }
 
-    public function resolveUser(): User
+    public static function resolveUser(): User
     {
         $user = new User();
         $user->name = 'foo';
 
         return $user;
+    }
+
+    public function testThrowsIfResolvedIsUsedOnMutation(): void
+    {
+        $this->expectExceptionObject(CanDirective::resolvedIsUnsafeInMutations('foo'));
+        $this->buildSchema(/** @lang GraphQL */ '
+        type Mutation {
+            foo: ID @can(resolved: true)
+        }
+        ');
     }
 }